language support, formatting, sorting, missing image
Changing the language from German to English is now simple, so is changing the author or title. Now only 4 entries per page are shown, but correctly spaced. 2x2 is converted to 2\times 2. Entries are sorted alphabetically by their brand and name (not longer by the filename of the image). It is now easily possible to only show an arbitrary combination of distinct cubes, shapemods or stickermods. A missing image now has a grey cube along with the text "no image found".